requirements.txt
requests
python-dotenv
twelvedata